refactor: TicketList
